* Fixed bug in Thumb2 pop caused by me incorrectly assuming that   ARM_SP == 13, ARM_LR == 14, and ARM_PC == 15, which is not the case * updated CMakeLists to include building arm regression test * added explicit casts for 64 bit visual studio 2012 build to get around   truncation warnings from size_t conversion

+++ SStream.c
@@ -24,7 +24,7 @@ void SStream_Init(SStream *ss)
 void SStream_concat0(SStream *ss, char *s)
 {
 #ifndef CAPSTONE_DIET
-	unsigned int len = strlen(s);
+	unsigned int len = (unsigned int) strlen(s);
 
 	memcpy(ss->buffer + ss->index, s, len);
 	ss->index += len;
